Implement a basic storm implementation and a host class
I did some quick digging and asked around on IRC for a plain and simple
python ORM which handles sqlite3.  I found storm, a canonical project,
which seems to do the job quiet nicely.  Its an ORM, and thats it.
It's not perfect, but looks like it'll do the job just fine.